Move mature APIs to include/xrt folder
