+++ src/runtime_src/core/include/experimental/CMakeLists.txt
@@ -15,9 +15,3 @@ set(XRT_EXPERIMENTAL_HEADER_SRC
   xclbin-util.h)
 
 install (FILES ${XRT_EXPERIMENTAL_HEADER_SRC} DESTINATION ${XRT_INSTALL_INCLUDE_DIR}/experimental COMPONENT ${XRT_DEV_COMPONENT})
-
-set(XRT_EXPERIMENTAL_DETAIL_HEADER_SRC
-  detail/param_traits.h
-  detail/pimpl.h)
-  
-install (FILES ${XRT_EXPERIMENTAL_DETAIL_HEADER_SRC} DESTINATION ${XRT_INSTALL_INCLUDE_DIR}/experimental/detail COMPONENT ${XRT_DEV_COMPONENT})
